fix: separate game test and game file
