+++ 06.bowling_object/game.rb
@@ -1,7 +1,4 @@
 require_relative 'frame'
-
-require 'minitest/autorun'
-require 'minitest/pride'
 require_relative 'bowling'
 
 class Game
@@ -47,35 +44,4 @@ end
 # assert_equal 164, game.calc_scores
 
 # # テストを追加
-
-class BowlingTest < Minitest::Test
-  def test_score1
-    marks = '6,3,9,0,0,3,8,2,7,3,X,9,1,8,0,X,6,4,5'
-    game = Game.new(marks)
-    assert_equal 139, game.calc_scores
-  end
-
-  def test_score2
-    marks = '6,3,9,0,0,3,8,2,7,3,X,9,1,8,0,X,X,X,X'
-    game = Game.new(marks)
-    assert_equal 164, game.calc_scores
-  end
-
-  def test_score3
-    marks = '0,10,1,5,0,0,0,0,X,X,X,5,1,8,1,0,4'
-    game = Game.new(marks)
-    assert_equal 107, game.calc_scores
-  end
-
-  def test_score4
-    marks = '6,3,9,0,0,3,8,2,7,3,X,9,1,8,0,X,X,0,0'
-    game = Game.new(marks)
-    assert_equal 134, game.calc_scores
-  end
-
-  def test_perfect_game
-    marks = 'X,X,X,X,X,X,X,X,X,X,X,X'
-    game = Game.new(marks)
-    assert_equal 300, game.calc_scores
-  end
 end
